Address misaligned load/store in ByteArray's float accessors.
This can trigger faults on some ARM hardware.

Appease UBSAN that we want 2's complement shifts in the push integer bytecode.

Setup CI for the sanitizers.

+++ vm/interpreter.cc
@@ -1587,7 +1587,7 @@ void Interpreter::Interpret() {
     }
     case 249: {
       uint8_t byte2 = *ip_++;
-      int8_t byte3 = *ip_++;
+      uintptr_t byte3 = static_cast<intptr_t>(static_cast<int8_t>(*ip_++));
       Push(SmallInteger::New((byte3 << 8) | byte2));
       break;
     }
